backend/requirements.txt
attrs==17.4.0
decorator==4.2.1
Django==1.11.1
django-environ==0.4.4
django-money==0.12.3
django-test-plus==1.0.18
djangorestframework==3.7.7
factory-boy==2.9.2
Faker==0.8.10
ipdb==0.10.3
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.1
nameparser==0.5.6
parso==0.1.1
pexpect==4.3.1
pickleshare==0.7.4
Pillow==5.0.0
pluggy==0.6.0
prompt-toolkit==1.0.15
psycopg2==2.7.3.2
ptyprocess==0.5.2
py==1.5.2
py-moneyed==0.7.0
Pygments==2.2.0
pytest==3.3.2
pytest-django==2.9.1
python-dateutil==2.6.1
pytz==2017.3
simplegeneric==0.8.1
six==1.11.0
text-unidecode==1.1
traitlets==4.3.2
wcwidth==0.1.7
